+++ src/ImportCss/ImportCss.mjs
@@ -4,7 +4,7 @@
 export class ImportCss {
     /**
      * @param {string} url
-     * @returns {Promise<CSSStyleSheet | HTMLStyleElement>}
+     * @returns {Promise<CSSStyleSheet>}
      * @abstract
      */
     import(url) { }